Revert "Re-land API package (#4793)"
This reverts commit c9d4e2a0c8f5cacebfe99e2022d051e42a46fb23.
