+++ common.tsconfig.json
@@ -2,6 +2,7 @@
   "$schema": "https://json.schemastore.org/tsconfig",
   "compilerOptions": {
     "alwaysStrict": true,
+    "composite": true,
     "esModuleInterop": true,
     "ignoreDeprecations": "5.0",
     "importsNotUsedAsValues": "preserve",
