+++ package.json
@@ -58,8 +58,7 @@
       "packages/central-server",
       "packages/meta-server",
       "packages/shared",
-      "packages/constants",
-      "packages/api-client"
+      "packages/constants"
     ],
     "nohoist": [
       "**/fbjs-scripts",
