--- packages/api-client/.swcrc
@@ -1,15 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/swcrc",
-  "exclude": "node_modules/",
-  "module": {
-    "type": "es6"
-  },
-  "jsc": {
-    "parser": {
-      "syntax": "ecmascript",
-      "jsx": true
-    },
-    "target": "esnext",
-    "keepClassNames": true
-  }
-}
