--- packages/api-client/src/index.js
@@ -1,7 +1,0 @@
-export {
-  AuthExpiredError,
-  ServerResponseError,
-  ServerUnavailableError,
-  VersionIncompatibleError,
-} from './errors';
-export { TamanuApi } from './TamanuApi';
