+++ packages/web/package.json
@@ -46,7 +46,6 @@
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.61",
     "@react-pdf/renderer": "^3.1.3",
-    "@tamanu/api-client": "*",
     "@tamanu/shared": "*",
     "@tanstack/react-query": "^4.0.10",
     "@tanstack/react-query-devtools": "^4.0.10",
